refactor: unify runner TOCTOU checker with internal/security
