+++ internal/security/toctou_test.go
@@ -14,8 +14,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	commontesting "github.com/isseis/go-safe-cmd-runner/internal/common/testutil"
-	"github.com/isseis/go-safe-cmd-runner/internal/groupmembership"
-	isec "github.com/isseis/go-safe-cmd-runner/internal/security"
 )
 
 // TestCollectTOCTOUCheckDirs verifies that directory collection covers all required
@@ -75,7 +73,7 @@ func TestCollectTOCTOUCheckDirs(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := isec.CollectTOCTOUCheckDirs(tt.verifyFiles, tt.commandPaths, tt.hashDir)
+			got := CollectTOCTOUCheckDirs(tt.verifyFiles, tt.commandPaths, tt.hashDir)
 
 			// Sort both slices for comparison
 			gotSet := make(map[string]struct{}, len(got))
@@ -102,12 +100,10 @@ func TestRunTOCTOUPermissionCheck_NoViolations(t *testing.T) {
 	err := os.Chmod(tmpDir, 0o755)
 	require.NoError(t, err)
 
-	gm := groupmembership.New()
-
-	v, err := NewValidator(nil, WithGroupMembership(gm))
+	v, err := NewDirectoryPermChecker()
 	require.NoError(t, err)
 
-	violations := isec.RunTOCTOUPermissionCheck(v, []string{tmpDir}, slog.Default())
+	violations := RunTOCTOUPermissionCheck(v, []string{tmpDir}, slog.Default())
 	assert.Empty(t, violations, "no violations expected for secure directory")
 }
 
@@ -124,15 +120,13 @@ func TestRunTOCTOUPermissionCheck_ViolationDetected(t *testing.T) {
 		_ = os.Chmod(tmpDir, 0o755)
 	})
 
-	gm := groupmembership.New()
-
-	v, err := NewValidator(nil, WithGroupMembership(gm))
+	v, err := NewDirectoryPermChecker()
 	require.NoError(t, err)
 
-	violations := isec.RunTOCTOUPermissionCheck(v, []string{tmpDir}, slog.Default())
+	violations := RunTOCTOUPermissionCheck(v, []string{tmpDir}, slog.Default())
 	require.Len(t, violations, 1, "expected exactly one violation for world-writable directory")
 	assert.Equal(t, filepath.Clean(tmpDir), violations[0].Path)
-	assert.True(t, errors.Is(violations[0].Err, isec.ErrInvalidDirPermissions), "violation error should be about directory permissions")
+	assert.True(t, errors.Is(violations[0].Err, ErrInvalidDirPermissions), "violation error should be about directory permissions")
 }
 
 // TestRunTOCTOUPermissionCheck_MultipleViolations verifies that multiple
@@ -150,23 +144,19 @@ func TestRunTOCTOUPermissionCheck_MultipleViolations(t *testing.T) {
 		})
 	}
 
-	gm := groupmembership.New()
-
-	v, err := NewValidator(nil, WithGroupMembership(gm))
+	v, err := NewDirectoryPermChecker()
 	require.NoError(t, err)
 
-	violations := isec.RunTOCTOUPermissionCheck(v, []string{dir1, dir2}, slog.Default())
+	violations := RunTOCTOUPermissionCheck(v, []string{dir1, dir2}, slog.Default())
 	assert.Len(t, violations, 2, "expected two violations")
 }
 
 // TestRunTOCTOUPermissionCheck_EmptyDirs verifies that an empty directory list
 // produces no violations.
 func TestRunTOCTOUPermissionCheck_EmptyDirs(t *testing.T) {
-	gm := groupmembership.New()
-
-	v, err := NewValidator(nil, WithGroupMembership(gm))
+	v, err := NewDirectoryPermChecker()
 	require.NoError(t, err)
 
-	violations := isec.RunTOCTOUPermissionCheck(v, []string{}, slog.Default())
+	violations := RunTOCTOUPermissionCheck(v, []string{}, slog.Default())
 	assert.Empty(t, violations)
 }
